==> R/00_all.qmd <==
---
title: "Run-all Quarto documents in folder"
output:
  html_document:
    output_dir: "../results/"
editor: visual
---

# Run-all quarto documents in this folder

This script is running the entire program, by running all the quarto documents in the folder and rendering them to .html.

```{r}
# Load libraries
library(tidyverse)
library(quarto)
library(fs)
```

Make sure you are in the right project path, you can check you project path by doing getwd() in the console

```{r}
# Specify the directories
current_dir <- path("R")
output_dir <- path("../results")

# Ensure the output directory exists
dir_create(output_dir)

# List all .qmd files in the R folder, excluding this file
qmd_files <- dir_ls( glob = "*.qmd") |>
  discard(~ basename(.x) == "00_all.qmd")  # Exclude 00_all.qmd

# Print the files being rendered
cat("Rendering the following .qmd files to:", output_dir, "\n")
print(qmd_files)

# Render each .qmd file and save the results in the output directory
qmd_files |>
  walk(~ {
    # Render the file
    rendered_file <- quarto_render(input = .x)
})
```

```{r}
# List all files in the `current_dir` directory
all_files <- dir_ls()  # This defines `all_files`

# Identify non .qmd and .R files
non_qmd_r_files <- all_files |>
  discard(~ str_detect(.x, "\\.(qmd|R)$"))


#Delete the folders
folders_to_delete <- non_qmd_r_files |>
  discard(~ str_detect(.x, "\\.html$"))

walk(folders_to_delete, ~ {
    dir_delete(.x) # Deletes the directory and its contents
})


# Move non .html files to the specific directory
result_files <- non_qmd_r_files |>
  discard(~ .x %in% folders_to_delete)

result_files |> walk(~ {
  target_path <- output_dir
  file_move(.x, output_dir)
})


```
