==> R/03_augment.qmd <==
---
title: "Augment data"
format:
  html:
    output_dir: "../results/"
editor: visual
---

## Load libraries:

```{r}
#| message: false
library(readr)
library(dplyr)
library(tidyverse)
```

## Loading the data:

```{r}
#| message: false 
data <- read_tsv("../data/02_dat_clean.tsv")
```

## Creating new variables:

```{r}
#Adding age groups:
data <- data |>
  mutate(Age_group = cut(Age, breaks = seq(10, 100, by = 10)))

#Relocating the agegroups to be place behinde age:
data <- data |>
  relocate(Age_group, .after = Age)

#Adding numerical stages and metastasis:
data <- data |>
  mutate(No_stage = case_when(Dukes_stage == "A" ~ 1,
                            Dukes_stage == "B" ~ 2,
                            Dukes_stage == "C" ~ 3,
                            Dukes_stage == "D" ~ 4),
         metastasis = case_when(No_stage <= 3 ~ 0,
                                No_stage > 3 ~ 1))

#Relocating the numerical stages to be place behinde age:
data <- data |>
  relocate(No_stage, .after = Dukes_stage) |> 
  relocate(metastasis, .after = No_stage)

#Adding a DFS_label column, which contains the data of DFS, in categorical format (relevant for analysis_3)
data <- data |>
  mutate(DFS_label = case_when(DFS_event == 0 ~ "No return of disease",
                               DFS_event == 1 ~ "Return of disease"), 
         .after = DFS_event)

## Save the data

write_tsv(data, "../data/03_dat_aug.tsv")
```
