==> R/05_analysis_1.qmd <==
---
title: "Principal component analysis"
format:
  html:
    output_dir: "../results/"
editor: visual
---

## Principal component analysis (PCA):

Load libraries:

```{r}
#| message: false 
library(readr)
library(tidyverse)
library(broom)
library(cowplot)
library(dplyr)
library(tidyr)

```

Load data:

```{r}
#| message: false 
data <- read_tsv("../data/03_dat_aug.tsv")
data <- data |>
  column_to_rownames(var = "ID_REF")
```

Prepare data for PCA and run PCA

```{r}
# Handle missing data (if any)
cleaned_data <- data |>
  select(where(is.numeric))|> 
  mutate(across(where(is.numeric), ~ ifelse(is.na(.), mean(., na.rm = TRUE), .)))

# Perform PCA
pca <- prcomp(cleaned_data, scale = TRUE)

pca_aug <- pca |>
  augment(data)

```

```{r}
scree_plot<- pca |>
  tidy("pcs") |> 
  mutate(percent = percent * 100) |> 
  ggplot(aes(x = PC,
             y = percent)) +
  geom_hline(yintercept = 0) +
  geom_col(colour = "black",
           alpha = 0.5) +
  theme_bw(base_size = 20) +
  theme(panel.grid.minor.x = element_blank(),
        panel.grid.major.x = element_blank()) +
  labs(title = "Scree Plot of PCA of Data",
       y = "Percent %")
scree_plot
#Save plot
scree_plot |> 
  ggsave(filename = "05_pca_scree_plot.png",
         path =  "../results/",
         bg = "white")


```

```{r}
pca_plot_axes_labels <- pca |>
  tidy("eigenvalues") |>
  mutate(lbl = str_c("PC", PC, ": VE = ", round(percent*100,2), "%")) |> 
  pull("lbl")
```

```{r}
#code plot
pca_plot <- pca_aug |> 
  ggplot(aes(x = .fittedPC1,
             y = .fittedPC2,
             colour = Dukes_stage,)) +
  geom_smooth(method = "lm") +
  geom_point() +
  coord_fixed() +
  labs(
    title = "PCA: Scores Plot of Colorectal Cancer",
    x = pluck(pca_plot_axes_labels, 1),
    y = pluck(pca_plot_axes_labels, 2),
    colour = "Duke stages") +
  theme_minimal()

#load plot
pca_plot

#Save plot
pca_plot |> 
  ggsave(filename = "05_pca_Dukes_stages.png",
         path =  "../results/",
         bg = "white")
```

```{r}
#code plot
pca_plot <- pca_aug |> 
  ggplot(aes(x = .fittedPC1,
             y = .fittedPC2,
             colour = Location,)) +
  geom_point() +  # Scatter plot
  geom_smooth(method = "lm") +
  geom_point() +
  coord_fixed() +
  labs(
    title = "PCA: Scores Plot of Colorectal Cancer",
    x = pluck(pca_plot_axes_labels, 1),
    y = pluck(pca_plot_axes_labels, 2),
    colour = "Location") +
  theme_minimal()

#load plot
pca_plot

#Save plot
pca_plot |> 
  ggsave(filename = "05_pca_location.png",
         path =  "../results/",
         bg = "white")
```

```{r}
#code plot
pca_plot <- pca_aug |> 
  ggplot(aes(x = .fittedPC1,
             y = .fittedPC2,
             colour = Gender,)) +
  geom_smooth(method = "lm") +
  geom_point() +
  coord_fixed() +
  labs(
    title = "PCA: Scores Plot of Colorectal Cancer",
    x = pluck(pca_plot_axes_labels, 1),
    y = pluck(pca_plot_axes_labels, 2),
    colour = "Location") +
  theme_minimal()

#load plot
pca_plot

#Save plot
pca_plot |> 
  ggsave(filename = "05_pca_gender.png",
         path =  "../results/",
         bg = "white")
```

## Clean environment

```{r}
rm(list = ls())
```
