==> R/06_analysis_2.qmd <==
---
title: "Up- or down-regulated genes"
format:
  html:
    output_dir: "../results"
editor: visual
---

## Aim

In this analysis, we will try to answer the following question

**What genes are significantly up- or down-regulated between the patients with/without metastasis?**

## Load library

```{r}
#| message: false 
library("tidyverse")
library("dplyr")
library("readr")
library("tibble")
library("broom")
library("ggrepel")
```

## Load data

```{r}
#| message: false 
data <- read_tsv("../data/03_dat_aug.tsv")
```

## Analysis

### Prepare the data

We pivot the data, so we can work *per gene*, then group per gene and nest the data.

```{r}
data_long_nested <- data |>
  select(ends_with("_at"), metastasis) |> 
  pivot_longer(cols = ends_with("_at"), 
               names_to = "gene", 
               values_to = "log2_expr_level") |> 
  group_by(gene) |> 
  nest() |> 
  ungroup()

```

### Fitting Models

We fit a linear model to each gene. Then, we tidy the fitted model, including getting confidence intervals.

```{r}
data_long_nested <- data_long_nested |> 
  group_by(gene) |> 
  mutate(model_object = map(.x = data,
                   .f = ~lm(formula = log2_expr_level ~ metastasis,
                            data = .x))) |> 
  ungroup() |> 
  mutate(model_object_tidy = map(.x = model_object,
                                 .f = ~tidy(x =.x,
                                            conf.int = TRUE,
                                            conf.level = 0.95)))
```

### Wrangling

We extract the model parameters, correct for multiple testing and add an indicator for significant findings

```{r}

estimates <- data_long_nested |> 
  unnest(model_object_tidy) |> 
  filter(term == "metastasis") |> 
  select("gene", "p.value", "estimate", "conf.low", "conf.high") |> 
  mutate(q.value = p.adjust(p.value),
         is_significant = case_when(q.value > 0.05 ~ "no",
                                    q.value <= 0.05 ~ "yes"))
```

### Inspection and approximation

```{r}
estimates |> 
  group_by(is_significant) |>  
  summarise(count = n())
```

When looking at the results after correction for multiple testing, we can see that no genes are significantly up- or down-regulated between patients with and without metastasis.

Therefore, for the remainder of our analysis, we will skip the correction for multiple testing and focus on the genes with the lowest p-values (\<0.01). This approach has significant limitations, as we risk detecting genes that appear to be significantly regulated purely by random chance. However, our goal here is simply to demonstrate our ability in data representation within the context of this course, rather than to obtain definitive or real results.

```{r}
estimates <- estimates |>  
  mutate(is_significant = case_when(p.value > 0.01 ~ "no",
                                    p.value <= 0.01 ~ "yes"))

estimates |> 
  group_by(is_significant) |> 
  summarise(count = n())
```

We will now consider that 20 genes out of the 1935 initial ones can be considered significant.

### Plots

```{r}
#| message: false 
pl1 <- estimates |> 
  filter(is_significant == "yes") |>
  mutate(gene = fct_reorder(gene, estimate)) |> 
  ggplot(aes(x = estimate,
             y = gene)) +
  geom_point() +
  geom_errorbarh(aes(xmax = conf.low, 
                     xmin = conf.high)) +
  labs(x = "Estimates (95%Cls)",
       y = "",
       title = "Genes Associated with Metastasis in Colorectal Cancer",
       caption = "Data from https://www.kaggle.com/datasets/amandam1/colorectal-cancer-patients/data") +
  theme_minimal(base_size = 10) +
  geom_vline(xintercept = 0)

# Save the plot
pl1 |> 
  ggsave(filename = "06_analysis_metastasis_regulation.png",
         path = "../results/",
         bg = "white",
         width = 5, 
         height = 4)

pl1
```

```{r}
#| message: false 
pl2 <- estimates |> 
  mutate(gen_sig = case_when(is_significant == "no" ~ "",
                             is_significant == "yes" ~ gene)) |> 
  ggplot(aes(x = estimate,
             y = -log10(p.value),
             color = is_significant,
             label = gen_sig)) +
  geom_text_repel(size = 3, 
                  max.overlaps = Inf) +
  geom_point(alpha = 0.5, 
             size = 1,
             show.legend = FALSE,
             shape = 18) +
  labs(x = "Estimates (95%Cls)",
       y = "-log10(p)",
       title = "Genes Associated with Metastasis in Colorectal Cancer",
       subtitle = "Genes highlighted in turuqoise were considered significant (but without multiple testing correction)",
       caption = "Data from https://www.kaggle.com/datasets/amandam1/colorectal-cancer-patients/data") +
  theme_minimal(base_size = 10) +
  theme(legend.position = "none",
        plot.subtitle = element_text(hjust = 1)) +
  geom_hline(yintercept = 0)

# save the plot
pl2 |> 
  ggsave(filename = "06_analysis_volcano.png",
         path ="../results/",
         bg = "white")

pl2
```

### Conclusion

Among the 1,935 genes analyzed, none were identified as significantly associated with the development of metastasis after multiple test correcting. However, out of the genes with the lowest p-value, 12 were found to be up-regulated, while 8 were found to be down-regulated in patients with metastasis compared to those without it.
