==> R/07_analysis_3.qmd <==
---
title: "Analysis 3"
output:
  html_document:
    output_dir: "../results/"
editor: visual
---

## Aim

In this analysis, we will try to answer the following question:

Is there a correlation between Disease-Free Survival and tumor location?

## Load libraries

```{r}
library(tidyverse)
library(dplyr)
library(ggplot2)
```

## Loading the data:

```{r}
data <- read_tsv("../data/03_dat_aug.tsv")
```

## Analysis

### Prepare and select the data

We start by grouping the samples based on their Disease Free Survival event, i.e. whether or not the disease showed a return or not. To make the grouping and visualization better, we first mutated the column to create a new, DFS_label column, containing the same data but in a categorical format.

After separating, we counted how many samples were in each location for each of the two groups. Then we created a column plot with ggplot2:

### Plotting

```{r}
dfs_loc_plot <- data |>
  group_by(DFS_label) |>
  count(Location) |>
  ggplot(mapping = aes(x = Location, y = n, fill=DFS_label)) +
  geom_col(position = "dodge", alpha=0.5, color="black") +
  theme_minimal() +
  theme(
    legend.position = "right",
    axis.text.x = element_text(angle = 10, vjust = 1, hjust=1),
    panel.grid.major.x = element_blank(),  # Remove major vertical grid lines for a cleaner graph
    panel.grid.minor.x = element_blank()   # Remove minor vertical grid lines for a cleaner graph
  ) +
  scale_y_continuous(breaks = seq(0, 20, by=2)) + # Show more ticks on the y axis to make the graph easier to read
  labs(
    title = "Distribution of Disease-Free Survival Events by Sample Location",
    subtitle = "Analysis of colorectal cancer samples",
    x = "Location of sample",
    y = "Number of samples",
    fill = NULL,
    caption = "Data: https://www.kaggle.com/datasets/amandam1/colorectal-cancer-patients/data"
  )

# Show plot
dfs_loc_plot
```

```{r}
# Save the plot
dfs_loc_plot |> 
  ggsave(filename = "07_analysis_3.png",
         path =  "../results/")
```

## Results

There is a clear distinction shown for the Rectum position. The majority of samples end up exhibiting a return of the disease, more than triple the amount. This is a bigger difference than any other location, making the rectum the most aggressive location. In contrast, no difference is shown for the colon location. The left side is also somewhat aggressive, while the right mostly shows no return of disease.
